[API] :hammer: export message and payload objects
